requirements.txt
speech_recognition>=3.10.0
pyttsx3>=2.90
requests>=2.31.0
gtts>=2.5.0
pygame>=2.5.0
google-genai>=0.8.0